requirements.txt
numpy==1.14.6
h5py==2.8.0
Pillow==5.2.0
tqdm==4.26.0
termcolor==1.1.0
